S20-579:refactor button component
