+++ src/components/Button/index.ts
@@ -1,4 +1,5 @@
-import { Button, ButtonProps } from "./Button";
+import { Button } from "./Button";
+import { ButtonProps } from "./Button.props";
 
 export { Button };
 export type { ButtonProps };
