+++ src/styled/Grid/Grid.props.ts
@@ -1,5 +1,5 @@
 import { GridProps, LayoutProps, PositionProps } from "styled-system";
 
-type GridComponentProps = GridProps & LayoutProps & PositionProps
+type GridComponentProps = GridProps & LayoutProps & PositionProps;
 
 export { GridComponentProps };
