+++ src/styled/Grid/index.ts
@@ -1,2 +1,1 @@
-
 export * from "./Grid";
